requirements.txt
# Web framework
fastapi==0.95.2
uvicorn[standard]==0.22.0

# Database and ORM
#asyncpg==0.27.0
SQLAlchemy==2.0.20
psycopg[binary]==3.1.12
databases
alembic==1.12.0

# Async HTTP and utilities
httpx==0.24.0

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation and Settings
pydantic==1.10.9
python-dotenv==1.0.0

# Caching / Redis
redis==4.5.5
aioredis==2.0.1

# Fuzzy matching
rapidfuzz==2.13.7

# Testing
pytest==7.4.0
pytest-asyncio==0.22.0
